Brief should identify review that has already happened.
I think this is helpful for identifying gaps.  For example, if the team
notes that they've talked to two other teams about their design, that
might prompt some of the reviewers (or the team itself, as they write
the brief) to notice that there's a third team that they also need to
talk to.